test(reactElementToJsxString): move owner test case to index.spec.js
